More work on Projects in Toolkit

+++ library/fui/fui_lcl_managers.pas
@@ -51,12 +51,12 @@ uses
   fhir_objects, fhir_factory, fhir_parser;
 
 type
-  TNodeOperation = (opAdd, opDelete, opEdit, opExecute, opOrder, opHeirarchy);
+  TNodeOperation = (opAdd, opDelete, opEdit, opExecute, opOrder, opHeirarchy, opRefresh);
   TNodeOperationSet = set of TNodeOperation;
 
   { TControlEntry }
 
-  TControlOperation = (copAdd, copEdit, copDelete, copUp, copDown, copReload, copExecute);
+  TControlOperation = (copNone, copAdd, copEdit, copDelete, copUp, copDown, copReload, copExecute, copRefresh);
 
   TControlEntry = class (TFslObject)
   private
@@ -65,6 +65,7 @@ type
     FOp: TControlOperation;
   public
     function link : TControlEntry; overload;
+
 
     property control : TControl read FControl write FControl;
     property op : TControlOperation read FOp write FOp;
@@ -87,16 +88,20 @@ type
 
     procedure doControl(sender : TObject); virtual; abstract;
     procedure doMnuClick(Sender: TObject); virtual; abstract;
+    procedure updateControls(op : TControlOperation; allowed : boolean);
   public
     constructor Create; override;
     destructor Destroy; override;
 
     property OnSetFocus : TNotifyEvent read FOnSetFocus write FOnSetFocus;
+    property Images : TImagelist read FImages write FImages;
 
     function doubleClickEdit : boolean; virtual;
+    function AskOnDelete : boolean; virtual;
 
     procedure registerControl(c : TControl; op : TControlOperation; mode : String = '');
-    procedure registerMenuEntry(caption : String; imageIndex : integer; op : TControlOperation; mode : String = '');
+    function registerMenuEntry(caption : String; imageIndex : integer; op : TControlOperation; mode : String = '') : TMenuItem;
+    function registerSubMenuEntry(parent : TMenuItem; caption : String; imageIndex : integer; op : TControlOperation; mode : String = '') : TMenuItem;
   end;
 
   TListManager<T : TFslObject> = class abstract (TListOrTreeManagerBase)
@@ -117,7 +122,6 @@ type
     procedure SetList(AValue: TListView);
     procedure SetSettings(AValue: TIniFile);
     procedure updateStatus;
-    procedure updateControls(op : TControlOperation; allowed : boolean);
     procedure doListCompare(Sender: TObject; Item1, Item2: TListItem; Data: Integer; var result: Integer);
     procedure doListChange(Sender: TObject; Item: TListItem; Selected: Boolean);
     procedure doListDoubleClick(Sender: TObject);
@@ -126,7 +130,6 @@ type
   protected
     procedure doControl(sender : TObject); override;
     procedure doMnuClick(Sender: TObject); override;
-    function doubleClickEdit : boolean; virtual;
   public
     constructor Create; override;
     destructor Destroy; override;
@@ -136,7 +139,6 @@ type
     property Focus : T read GetFocus;
     property hasFocus : boolean read GetHasFocus;
     property List : TListView read FList write SetList;
-    property Images : TImagelist read FImages write FImages;
     property Filter : TEdit read FFilter write SetFilter;
     Property Enabled : boolean read FEnabled write SetEnabled;
     property Settings : TIniFile read FSettings write SetSettings;
@@ -170,11 +172,14 @@ type
 
     function addItem(mode : String) : T; virtual;
     function editItem(item : T; mode : String) : boolean; virtual;
-    procedure deleteItem(item : T); virtual;
+    function deleteItem(item : T) : boolean; virtual;
     function executeItem(item : T; mode : String) : boolean; virtual;
+    function refreshItem(item : T) : boolean; virtual;
   end;
 
   TFslTreeNode = class abstract (TFslObject)
+  private
+    FNode : TTreeNode;
   protected
     function getChildCount : integer; virtual; abstract;
     function getChild(index : integer) : TFslTreeNode; virtual; abstract;
@@ -192,11 +197,21 @@ type
     procedure SetSettings(AValue: TIniFile);
     procedure SetEnabled(AValue: boolean);
 
+    procedure clearTreeNode(item : T);
     procedure buildTreeNode(parent : TTreenode; item : T);
+    procedure populateTreeNode(item : T);
+    procedure refreshTreeNode(item : T);
     procedure LoadTreeNodes;
+    procedure doTreeChange(Sender: TObject);
+    procedure doTreeDoubleClick(Sender: TObject);
+    procedure updateStatus;
+    procedure DoEdited(Sender: TObject; Node: TTreeNode; var S: string);
   protected
     FData : TFslList<T>; // hidden root
 
+    procedure doControl(sender : TObject); override;
+    procedure doMnuClick(Sender: TObject); override;
+    function readOnly : boolean; virtual;
   public
     constructor Create; override;
     destructor Destroy; override;
@@ -213,6 +228,12 @@ type
     procedure refresh(item : T = nil); // leaves the items in place (and doesn't refilter) but updates text displays. it item = nil, updates all displayed items
     procedure saveStatus;
 
+    // control. These are not usually needed from outside
+    procedure doAdd(mode : String);
+    procedure doEdit(mode : String);
+    procedure doDelete(mode : String);
+    procedure doExecute(mode : String);
+
     // to override:
     function LoadData : boolean; virtual; abstract;
     function allowedOperations(item : T) : TNodeOperationSet; virtual; abstract; // return what is allowed in principle; no need to be concerned with the selection, except for whether modify/delete is allowed
@@ -223,10 +244,12 @@ type
     function getCellColors(item : T; var fore, back : TColor) : boolean; virtual;
     function getSummaryText(item : T) : String; virtual;
 
-    function addItem(mode : String) : T; virtual;
+    function addItem(parent : T; mode : String) : T; virtual;
     function editItem(item : T; mode : String) : boolean; virtual;
-    procedure deleteItem(item : T); virtual;
+    function editItemText(parent, item : T; var text : String) : boolean; virtual;
+    function deleteItem(parent, item : T) : boolean; virtual; // parent might be nil if we're at the root
     function executeItem(item : T; mode : String) : boolean; virtual;
+    function refreshItem(item : T) : boolean; virtual;
   end;
 
 //  public
@@ -501,15 +524,22 @@ constructor TListOrTreeManagerBase.Create;
 begin
   inherited Create;
   FPopup := TPopupMenu.create(nil);
+  FControls := TFslList<TControlEntry>.create;
 end;
 
 destructor TListOrTreeManagerBase.Destroy;
 begin
   FPopUp.Free;
+  FControls.Free;
   inherited Destroy;
 end;
 
 function TListOrTreeManagerBase.doubleClickEdit: boolean;
+begin
+  result := true;
+end;
+
+function TListOrTreeManagerBase.AskOnDelete: boolean;
 begin
   result := true;
 end;
@@ -531,18 +561,43 @@ begin
   end;
 end;
 
-procedure TListOrTreeManagerBase.registerMenuEntry(caption: String; imageIndex: integer; op: TControlOperation; mode : String = '');
-var
-  item : TMenuItem;
+function TListOrTreeManagerBase.registerMenuEntry(caption: String; imageIndex: integer; op: TControlOperation; mode : String = '') : TMenuItem;
 begin
-  item  := TMenuItem.create(nil);
-  FPopup.Items.add(item);
-  item.caption := caption;
-  item.imageIndex := imageIndex;
-  item.Tag := Integer(op);
+  result  := TMenuItem.create(nil);
+  FPopup.Items.add(result);
+  result.caption := caption;
+  result.imageIndex := imageIndex;
+  result.Tag := Integer(op);
   if mode <> '' then
-    item.name := 'mnuMode'+mode;
-  item.OnClick := doMnuClick;
+    result.name := 'mnuMode'+mode;
+  if op <> copNone then
+    result.OnClick := doMnuClick;
+end;
+
+function TListOrTreeManagerBase.registerSubMenuEntry(parent: TMenuItem; caption: String; imageIndex: integer; op: TControlOperation; mode: String): TMenuItem;
+begin
+  result  := TMenuItem.create(nil);
+  parent.add(result);
+  result.caption := caption;
+  result.imageIndex := imageIndex;
+  result.Tag := Integer(op);
+  if mode <> '' then
+    result.name := 'mnuMode'+mode;
+  result.OnClick := doMnuClick;
+end;
+
+procedure TListOrTreeManagerBase.updateControls(op: TControlOperation; allowed: boolean);
+var
+  entry : TControlEntry;
+  mnu : TMenuItem;
+  i : integer;
+begin
+  for entry in FControls do
+    if entry.op = op then
+      entry.control.enabled := allowed;
+  for i := 0 to FPopup.Items.Count - 1 do
+    if (TControlOperation(FPopup.Items[i].tag) = op) then
+      FPopup.Items[i].enabled := allowed;
 end;
 
 { TListManager }
@@ -552,12 +607,10 @@ begin
   inherited Create;
   FData := TFslList<T>.create;
   FFiltered := TFslList<T>.create;
-  FControls := TFslList<TControlEntry>.create;
 end;
 
 destructor TListManager<T>.Destroy;
 begin
-  FControls.Free;
   FFiltered.free;
   FData.free;
   inherited Destroy;
@@ -602,6 +655,7 @@ begin
         copUp : doUp;
         copDown : doDown;
         copReload : doLoad;
+        copRefresh : refresh(focus);
         copExecute : doExecute(entry.mode);
       end;
 end;
@@ -707,26 +761,12 @@ begin
   updateControls(copDelete, (opDelete in ops) and (i > -1));
   updateControls(copUp, (opOrder in ops) and (i > 0) and not Filtered);
   updateControls(copDown, (opOrder in ops) and (i > -1) and (i < FFiltered.count - 1) and not Filtered);
-  updateControls(copReload, true);
+  updateControls(copRefresh, opRefresh in ops);
   updateControls(copExecute, opExecute in ops);
   FCanEdit := opEdit in ops;
 
   if assigned(FOnSetFocus) then
     FOnSetFocus(self);
-end;
-
-procedure TListManager<T>.updateControls(op: TControlOperation; allowed: boolean);
-var
-  entry : TControlEntry;
-  mnu : TMenuItem;
-  i : integer;
-begin
-  for entry in FControls do
-    if entry.op = op then
-      entry.control.enabled := allowed;
-  for i := 0 to FPopup.Items.Count - 1 do
-    if (TControlOperation(FPopup.Items[i].tag) = op) then
-      FPopup.Items[i].enabled := allowed;
 end;
 
 procedure TListManager<T>.doListCompare(Sender: TObject; Item1, Item2: TListItem; Data: Integer; var result: Integer);
@@ -797,6 +837,7 @@ begin
     copUp : doUp;
     copDown : doDown;
     copReload : doLoad;
+    copRefresh : refresh(focus);
     copExecute : doExecute(mode);
   end;
 end;
@@ -888,13 +929,26 @@ var
   i : integer;
 begin
   f := focus;
-  if (QuestionDlg('Delete', 'Delete '+getSummaryText(f)+'?', mtConfirmation, [mrYes, mrNo], '') = mrYes) then
+  if AskOnDelete then
   begin
-    DeleteItem(f);
-    i := FList.itemIndex;
-    FData.Delete(i);
-    FList.Items.Delete(i);
-    updateStatus;
+    if (QuestionDlg('Delete', 'Delete '+getSummaryText(f)+'?', mtConfirmation, [mrYes, mrNo], '') = mrYes) then
+    begin
+      DeleteItem(f);
+      i := FList.itemIndex;
+      FData.Delete(i);
+      FList.Items.Delete(i);
+      updateStatus;
+    end;
+  end
+  else
+  begin
+    if DeleteItem(f) then
+    begin
+      i := FList.itemIndex;
+      FData.Delete(i);
+      FList.Items.Delete(i);
+      updateStatus;
+    end;
   end;
 end;
 
@@ -970,6 +1024,9 @@ var
   entry : TListItem;
   i, c : integer;
 begin
+  if (item <> nil) then
+    refreshItem(item);
+
   FList.BeginUpdate;
   try
     for entry in FList.items do
@@ -986,11 +1043,6 @@ end;
 function TListManager<T>.canSort: boolean;
 begin
   result := false;
-end;
-
-function TListManager<T>.doubleClickEdit: boolean;
-begin
-  result := true;
 end;
 
 function TListManager<T>.getCellColors(item: T; col: integer; var fore, back: TColor): boolean;
@@ -1023,7 +1075,7 @@ begin
   result := false;
 end;
 
-procedure TListManager<T>.deleteItem(item: T);
+function TListManager<T>.deleteItem(item: T) : boolean;
 begin
   raise Exception.create('Delete is not supported here');
 end;
@@ -1031,6 +1083,11 @@ end;
 function TListManager<T>.executeItem(item: T; mode: String): boolean;
 begin
   raise Exception.create('Execute is not supported here');
+end;
+
+function TListManager<T>.refreshItem(item: T) : boolean;
+begin
+  result := false;
 end;
 
 
@@ -1316,13 +1373,93 @@ begin
 end;
 
 procedure TTreeManager<T>.refresh(item: T);
+var
+  i : T;
 begin
-
+  if (item = nil) then
+  begin
+    for i in FData do
+      refreshTreeNode(i)
+  end
+  else if not refreshItem(item) then
+    refreshTreeNode(item);
 end;
 
 procedure TTreeManager<T>.saveStatus;
 begin
+  // no status to save
+end;
 
+procedure TTreeManager<T>.doAdd(mode: String);
+var
+  item, pp : T;
+  p : TTreeNode;
+  s : String;
+  i : integer;
+begin
+  pp := focus;
+  if pp = nil then
+    p := nil
+  else
+    p := pp.FNode;
+  item := addItem(pp, mode);
+  if item <> nil then
+  begin
+    if (p = nil) then
+      item.FNode := FTree.Items.add(nil, '')
+    else
+      item.FNode := FTree.Items.addChild(p, '');
+    item.FNode.Data := pointer(item);
+    populateTreeNode(item);
+    FTree.Selected := item.FNode;
+    updateStatus;
+  end;
+end;
+
+procedure TTreeManager<T>.doEdit(mode: String);
+begin
+  raise ETodo.create('doEdit');
+end;
+
+procedure TTreeManager<T>.doDelete(mode: String);
+var
+  f, pp : T;
+  n, p : TTreeNode;
+  s : String;
+  i : integer;
+begin
+  f := focus;
+  n := f.FNode;
+  p := n.parent;
+  if (p = nil) then
+    pp := nil
+  else
+    pp := TFslObject(p.Data) as T;
+
+  if AskOnDelete then
+  begin
+    if (QuestionDlg('Delete', 'Delete '+getSummaryText(f)+' from '+getSummaryText(pp)+'?', mtConfirmation, [mrYes, mrNo], '') = mrYes) then
+    begin
+      DeleteItem(pp, f);
+      FTree.items.delete(n);
+      updateStatus;
+    end;
+  end
+  else
+  begin
+    if DeleteItem(pp, f) then
+    begin
+      FTree.items.delete(n);
+      updateStatus;
+    end;
+  end;
+end;
+
+procedure TTreeManager<T>.doExecute(mode: String);
+begin
+  if HasFocus then
+    if ExecuteItem(getFocus, mode) then
+      refreshTreeNode(getFocus);
 end;
 
 procedure TTreeManager<T>.setTree(value: TTreeView);
@@ -1330,22 +1467,15 @@ var
   i : integer;
 begin
   FTree := value;
-  //List.AutoWidthLastColumn := true;
-  //List.RowSelect := true;
-  //List.OnCompare := doListCompare;
-  //List.OnSelectItem := doListChange;
-  //List.OnDblClick := doListDoubleClick;
-  //List.OnCustomDrawSubItem := doListDrawSubItem;
+  FTree.MultiSelect := false;
+  FTree.OnClick := doTreeChange;
+  Tree.OnDblClick := doTreeDoubleClick;
   FPopup.Images := FImages;
   buildMenu;
   if FPopup.Items.Count > 0 then
     FTree.PopupMenu := FPopup;
-  //if FSettings <> nil then
-  //  for i := 0 to List.columns.count - 1 do
-  //    list.columns[i].width := FSettings.ReadInteger(List.Name, 'column'+inttostr(i), list.columns[i].width);
-  FTree.ReadOnly := true;
-  //FTree.AutoSort := false;
-  //FTree.AutoSortIndicator := false;
+  FTree.ReadOnly := readOnly;
+  FTree.OnEdited := DoEdited;
   FTree.SortType := stNone;
 end;
 
@@ -1357,12 +1487,14 @@ begin
 end;
 
 function TTreeManager<T>.getFocus : T;
+var
+  sel : TTreeNode;
 begin
-  //if (FList.itemindex = -1) or (FList.ItemIndex >= FFiltered.count) then
-  //  result := nil
-  //else
-  //  result := FFiltered[FList.itemIndex];
-  result := nil;
+  sel := Tree.Selected;
+  if sel = nil then
+    result := nil
+  else
+    result := TFslObject(sel.Data) as T;
 end;
 
 function TTreeManager<T>.GetHasFocus : boolean;
@@ -1387,16 +1519,48 @@ begin
 end;
 
 procedure TTreeManager<T>.buildTreeNode(parent : TTreenode; item : T);
+begin
+  if item.FNode <> nil then
+    raise ELibraryException.create('Node appears in the tree more than once');
+  if (parent = nil) then
+    item.FNode := FTree.Items.add(nil, '')
+  else
+    item.FNode := FTree.Items.addChild(parent, '');
+  item.FNode.Data := pointer(item);
+  populateTreeNode(item);
+end;
+
+procedure TTreeManager<T>.populateTreeNode(item : T);
 var
-  n : TTreeNode;
   i : integer;
 begin
-  n := FTree.Items.add(nil, getCellText(item));
+  item.FNode.Text := getCellText(item);
   i := getImageIndex(item);
-  n.ImageIndex := i;
-  n.SelectedIndex := i;
+  item.FNode.ImageIndex := i;
+  item.FNode.SelectedIndex := i;
   for i := 0 to item.getchildCount - 1 do
-    buildTreeNode(n, item.getChild(i) as T);
+    buildTreeNode(item.FNode, item.getChild(i) as T);
+end;
+
+procedure TTreeManager<T>.refreshTreeNode(item: T);
+var
+  i : integer;
+begin
+  item.FNode.Text := getCellText(item);
+  i := getImageIndex(item);
+  item.FNode.ImageIndex := i;
+  item.FNode.SelectedIndex := i;
+  for i := 0 to item.getchildCount - 1 do
+    refreshTreeNode(item.getChild(i) as T);
+end;
+
+procedure TTreeManager<T>.clearTreeNode(item : T);
+var
+  i : integer;
+begin
+  item.FNode := nil;
+  for i := 0 to item.getchildCount - 1 do
+    clearTreeNode(item.getChild(i) as T);
 end;
 
 procedure TTreeManager<T>.LoadTreeNodes;
@@ -1404,7 +1568,121 @@ var
   i : T;
 begin
   for i in FData do
+    clearTreeNode(i);
+
+  for i in FData do
     buildTreeNode(nil, i);
+end;
+
+procedure TTreeManager<T>.doTreeChange(Sender: TObject);
+begin
+  updateStatus;
+end;
+
+procedure TTreeManager<T>.doTreeDoubleClick(Sender: TObject);
+var
+  item : T;
+  i : integer;
+begin
+  if hasFocus then
+  begin
+    item := GetFocus;
+    if doubleClickEdit then
+    begin
+      if FCanEdit and hasFocus and EditItem(item, '') then
+      begin
+        for i := item.FNode.Count - 1 downto 0 do
+          item.FNode.Items[i].delete;
+        populateTreeNode(item);
+      end;
+    end
+    else
+      doExecute('');
+  end;
+end;
+
+procedure TTreeManager<T>.updateStatus;
+var
+  sel : TTreeNode;
+  ops : TNodeOperationSet;
+  op : TControlOperation;
+begin
+  sel := FTree.Selected;
+  if sel = nil then
+    ops := allowedOperations(nil)
+  else
+    ops := allowedOperations(TFslObject(sel.Data) as T);
+
+  updateControls(copAdd, opAdd in ops);
+  updateControls(copEdit, (opEdit in ops) and (sel <> nil));
+  updateControls(copDelete, (opDelete in ops) and (sel <> nil));
+  updateControls(copRefresh, opRefresh in ops);
+  updateControls(copExecute, opExecute in ops);
+  FCanEdit := opEdit in ops;
+
+  if assigned(FOnSetFocus) then
+    FOnSetFocus(self);
+end;
+
+procedure TTreeManager<T>.DoEdited(Sender: TObject; Node: TTreeNode; var S: string);
+var
+  item, pp : T;
+  p : TTreeNode;
+begin
+  item := focus;
+  p := item.FNode.parent;
+  if (p <> nil) then
+    pp := TFslObject(p.data) as T
+  else
+    pp := nil;
+  if not editItemText(pp, item, s) then
+    s := getCellText(item);
+end;
+
+procedure TTreeManager<T>.doControl(sender: TObject);
+var
+  entry : TControlEntry;
+begin
+  for entry in FControls do
+    if entry.control = sender then
+      case entry.op of
+        copAdd : doAdd(entry.mode);
+        copEdit : doEdit(entry.mode);
+        copDelete : doDelete(entry.mode);
+        //copUp : doUp;
+        //copDown : doDown;
+        copReload : doLoad;
+        copRefresh : refresh(focus);
+        copExecute : doExecute(entry.mode);
+      end;
+end;
+
+procedure TTreeManager<T>.doMnuClick(Sender: TObject);
+var
+  mnu : TMenuItem;
+  mode : String;
+begin
+  mnu := (Sender as TMenuItem);
+  if mnu.Name.StartsWith('mnuMode') then
+    mode := mnu.Name.Substring(7)
+  else
+    mode := '';
+
+  case TControlOperation(mnu.Tag) of
+    copAdd : doAdd(mode);
+    copEdit : doEdit(mode);
+    copDelete : doDelete(mode);
+    //copUp : doUp;
+    //copDown : doDown;
+    copReload : doLoad;
+    copRefresh : refresh(focus);
+    copExecute : doExecute(mode);
+  end;
+end;
+
+function TTreeManager<T>.readOnly: boolean;
+begin
+
 end;
 
 procedure TTreeManager<T>.buildMenu;
@@ -1426,7 +1704,7 @@ begin
   result := getCellText(item);
 end;
 
-function TTreeManager<T>.addItem(mode : String) : T;
+function TTreeManager<T>.addItem(parent : T; mode : String) : T;
 begin
   result := nil;
 end;
@@ -1436,7 +1714,12 @@ begin
   result := false;
 end;
 
-procedure TTreeManager<T>.deleteItem(item : T);
+function TTreeManager<T>.editItemText(parent, item: T; var text: String): boolean;
+begin
+    result := false;
+end;
+
+function TTreeManager<T>.deleteItem(parent, item : T) : boolean;
 begin
   raise Exception.create('Delete is not supported here');
 end;
@@ -1444,6 +1727,11 @@ end;
 function TTreeManager<T>.executeItem(item : T; mode : String) : boolean;
 begin
   raise Exception.create('Execute is not supported here');
+end;
+
+function TTreeManager<T>.refreshItem(item: T) : boolean;
+begin
+  result := false;
 end;
 
 End.
